# src/main/resources/openapi/distribution-spec.yaml
openapi: 3.0.3
info:
  version: "2.0"
  title: EuphoriaV docker registry API
  description: Docker registry API based on OCI Distribution Specification
servers:
  - description: Local server
    url: http://localhost:5000
...
paths:
...
  /v2/{name}/blobs/uploads/:
    post:
      tags:
        - Blobs
      summary: Initiate blob upload
      operationId: InitiateBlobUpload
      description: Initiate an upload session for a blob (layer or config) in a repository.
      parameters:
        - name: name
          in: path
          required: true
          description: Name of the target repository
          schema:
            type: string

      responses:
        "202":
          description: Upload initiated successfully.
          headers:
            Location:
              description: Upload location URL for `PATCH` or `PUT` requests
              schema:
                type: string
            Docker-Upload-UUID:
              description: Upload session UUID
              schema:
                type: string
            Range:
              description: Byte range uploaded so far (inclusive)
              schema:
                type: string
        "401":
          description: Authentication required.
        "403":
          description: Access denied.
        "429":
          description: Too many requests.
  /v2/{name}/blobs/{digest}:
    head:
      tags:
        - Blobs
      summary: Check existence of blob
      operationId: CheckBlobExists
      description: Check whether a blob (layer or config) exists in the registry.
      parameters:
        - name: name
          in: path
          required: true
          description: Name of the Repository
          schema:
            type: string
        - name: digest
          in: path
          required: true
          description: Digest of the blob
          schema:
            type: string

      responses:
        "200":
          description: Blob exists
          headers:
            Content-Length:
              description: Size of the blob in bytes
              schema:
                type: integer
            Docker-Content-Digest:
              description: Digest of the blob
              schema:
                type: string
            Content-Type:
              description: MIME type of the blob content
              schema:
                type: string
        "404":
          description: Blob not found
        "401":
          description: Authentication required
        "403":
          description: Access denied
        "429":
          description: Too many requests
    get:
      tags:
        - Blobs
      summary: Retrieve blob
      operationId: GetBlob
      description: |
        Download the blob identified by digest from the registry.

        Blobs include image layers and configuration objects. Clients must use the digest from the manifest to retrieve a blob.

        This endpoint may return a `307 Temporary Redirect` to a CDN or storage location. Clients must follow the redirect to obtain the actual blob content.

        The blob content is typically a gzipped tarball (for layers) or JSON (for configs). The MIME type is usually `application/octet-stream`.
      x-codeSamples:
        - lang: Bash
          label: cURL
          source: |
            # GET (download) a blob
            curl -L \
              -H "Authorization: Bearer $TOKEN" \
              https://registry-1.docker.io/v2/library/ubuntu/blobs/sha256:abc123... \
              -o layer.tar.gz
      parameters:
        - name: name
          in: path
          required: true
          description: Repository Name
          example: library/ubuntu
          schema:
            type: string
        - name: digest
          in: path
          required: true
          description: Digest of the Blob
          schema:
            type: string
          example: sha256:abc123def456...

      responses:
        "200":
          description: Blob content returned directly
          headers:
            Content-Length:
              description: Size of the blob in bytes
              schema:
                type: integer
              example: 32768
            Content-Type:
              description: MIME type of the blob
              schema:
                type: string
              example: application/octet-stream
            Docker-Content-Digest:
              description: Digest of the returned blob
              schema:
                type: string
              example: sha256:abc123def456...
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
              examples:
                small-layer:
                  summary: Example binary blob (gzipped tar layer)
                  value: "<binary data not shown>"

        "307":
          description: Temporary redirect to blob location
          headers:
            Location:
              description: Redirect URL for blob download (e.g., S3 or CDN)
              schema:
                type: string
              example: https://cdn.docker.io/blobs/library/ubuntu/abc123...
        "401":
          description: Authentication required
        "403":
          description: Access denied
        "404":
          description: Blob not found
        "429":
          description: Too many requests
  /v2/{name}/blobs/uploads/{uuid}:
    put:
      tags:
        - Blobs
      summary: Complete blob upload
      operationId: CompleteBlobUpload
      description: Complete the upload of a blob by finalizing an upload session.
      parameters:
        - name: name
          in: path
          required: true
          description: Repository name
          schema:
            type: string
        - name: uuid
          in: path
          required: true
          description: Upload session UUID returned from the POST request
          schema:
            type: string
            format: uuid
        - name: digest
          in: query
          required: true
          description: Digest of the uploaded blob
          schema:
            type: string
        - name: Content-Range
          in: header
          required: false
          schema:
            type: string
          description: Optional. Byte range of the chunk being sent
      requestBody:
        required: false
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary

      responses:
        "201":
          description: Upload completed successfully
          headers:
            Docker-Content-Digest:
              description: Canonical digest of the stored blob
              schema:
                type: string
            Location:
              description: URL where the blob is now accessible
              schema:
                type: string
            Content-Length:
              description: Always zero for completed uploads
              schema:
                type: integer
        "400":
          description: Invalid digest or missing parameters
        "401":
          description: Authentication required
        "403":
          description: Access denied
        "404":
          description: Upload session not found
        "416":
          description: Requested range not satisfiable (if used in chunked mode)
        "429":
          description: Too many requests

    patch:
      tags:
        - Blobs
      summary: Upload blob chunk
      operationId: UploadBlobChunk
      description: Upload a chunk of a blob to an active upload session.
      parameters:
        - name: name
          in: path
          required: true
          description: Repository name
          schema:
            type: string
        - name: uuid
          in: path
          required: true
          description: Upload session UUID
          schema:
            type: string
            format: uuid
        - name: Content-Range
          in: header
          required: false
          schema:
            type: string
          description: Optional. Byte range of the chunk being sent

      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary

      responses:
        "202":
          description: Chunk accepted and stored
          headers:
            Location:
              description: URL to continue or finalize the upload
              schema:
                type: string
            Range:
              description: Byte range uploaded so far (inclusive)
              schema:
                type: string
            Docker-Upload-UUID:
              description: Upload session UUID
              schema:
                type: string
        "400":
          description: Malformed content or range
        "401":
          description: Authentication required
        "403":
          description: Access denied
        "404":
          description: Upload session not found
        "416":
          description: Range error (e.g., chunk out of order)
        "429":
          description: Too many requests
    delete:
      tags:
        - Blobs
      summary: Cancel blob upload
      operationId: CancelBlobUpload
      description: Cancel an in-progress blob upload session.
      parameters:
        - name: name
          in: path
          required: true
          description: Name of the repository
          schema:
            type: string
        - name: uuid
          in: path
          required: true
          description: Upload session UUID
          schema:
            type: string
            format: uuid

      responses:
        "204":
          description: Upload session cancelled successfully. No body is returned.
        "401":
          description: Authentication required
        "403":
          description: Access denied
        "404":
          description: Upload session not found
        "429":
          description: Too many requests
...